requirements.txt

streamlit==1.28.0
python-docx==0.8.11
pdfplumber==0.10.3
spacy==3.7.2
nltk==3.8.1
anthropic==0.7.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
reportlab==4.0.7
requests==2.31.0
textblob==0.17.1
pydantic==2.5.0
unidecode==1.3.0
